+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
+- Support for Top Hotels checklist
 - Expandable visited map viewable when tapped
 
 ## [Version 1.2.0, Build 644] - 2019-10-01